[tune/structure] Move AutoML board
