+++ python/ray/tune/automl/board/frontend/view.py
@@ -1,10 +1,10 @@
 from django.shortcuts import render
 
-from ray.tune.automlboard.settings import (
+from ray.tune.automl.board.settings import (
     AUTOMLBOARD_RELOAD_INTERVAL,
     AUTOMLBOARD_LOG_DIR,
 )
-from ray.tune.automlboard.models.models import JobRecord, TrialRecord, ResultRecord
+from ray.tune.automl.board.models.models import JobRecord, TrialRecord, ResultRecord
 from ray.tune.trial import Trial
 
 import datetime
