+++ python/ray/tune/automl/board/manage.py
@@ -4,5 +4,5 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ray.tune.automlboard.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ray.tune.board.settings")
     execute_from_command_line(sys.argv)
