+++ python/ray/tune/automl/board/models/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "ray.tune.board.models.apps.ModelConfig"
