+++ python/ray/tune/automl/board/models/apps.py
@@ -4,4 +4,4 @@ from django.apps import AppConfig
 class ModelConfig(AppConfig):
     """Model Congig for models."""
 
-    name = "ray.tune.automlboard.models"
+    name = "ray.tune.board.models"
