--- python/ray/tune/automlboard/models/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = "ray.tune.automlboard.models.apps.ModelConfig"
